pyproject.toml
[project]
name = "flowchart-evaluator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.5",
    "numpy>=2.3.5",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "pytesseract>=0.3.13",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]
